move broken tests from models
